Atualização Automática via bash shell

+++ src/vps/compilar_vps.sh
@@ -283,27 +283,31 @@ then
 
 fi
 
-vagrant destroy -f
-sleep 3
-vagrant box remove neoricalex/nfdos
-sleep 3
-virsh vol-delete --pool default neoricalex-VAGRANTSLASH-nfdos_vagrant_box_image_0.img
-sleep 3
-echo "==> [DEBUG] vagrant global-status --prune"
-vagrant global-status --prune
-sleep 3
-echo "==> [DEBUG] vboxmanage list vms"
-vboxmanage list vms
-sleep 3
-echo "==> [DEBUG] vagrant box list"
-vagrant box list
-sleep 3
-echo "==> [DEBUG] virsh vol-list default"
-virsh vol-list default
-sleep 3
-
-#echo "Compilando o NFDOS..."
-#make nfdos
+limpar="1"
+if [ "$limpar" != "0" ]; then
+	echo "Limpando..."
+	vagrant destroy -f
+	sleep 3
+	vagrant box remove neoricalex/nfdos
+	sleep 3
+	virsh vol-delete --pool default neoricalex-VAGRANTSLASH-nfdos_vagrant_box_image_0.img
+	sleep 3
+	echo "==> [DEBUG] vagrant global-status --prune"
+	vagrant global-status --prune
+	sleep 3
+	echo "==> [DEBUG] vboxmanage list vms"
+	vboxmanage list vms
+	sleep 3
+	echo "==> [DEBUG] vagrant box list"
+	vagrant box list
+	sleep 3
+	echo "==> [DEBUG] virsh vol-list default"
+	virsh vol-list default
+	sleep 3
+else
+	echo "Compilando o NFDOS..."
+	make nfdos
+fi
 
 cd ..
 EOF
